API 구조에 맞게 변경사항 적용 / cart 포인트 충전 고장

+++ reportView_004/app/src/main/java/com/example/reportview_003/model/board/BoardCommentRequest.kt
@@ -2,7 +2,7 @@ package com.example.reportview_003.model.board
 
 
 data class BoardCommentRequest(
-    val qna_id: Int,
+    val qna_id: Long,
     val writer: String,
     val content: String
 )
