+++ reportView_004/app/src/main/java/com/example/reportview_003/model/purchase/CartGetItemRequest.kt
@@ -2,5 +2,5 @@ package com.example.reportview_003.model.purchase
 
 data class CartGetItemRequest(
     val user_id: String,
-    val book_id: Int
+    val book_id: Long
 )
